requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
google-genai>=0.2.0

